requirements.txt
streamlit==1.36.0
Pillow
moviepy
pygments
numpy